+++ YIF.Core.Domain/ServiceInterfaces/IIoEAdminService.cs
@@ -6,6 +6,7 @@ namespace YIF.Core.Domain.ServiceInterfaces
 {
     public interface IIoEAdminService
     {
+        Task<ResponseApiModel<DescriptionResponseApiModel>> ModifyDescriptionOfInstitution(string userId, InstitutionOfEducationPostApiModel institutionOfEducationPostApiModel);
         Task<ResponseApiModel<DescriptionResponseApiModel>> AddSpecialtyToIoe(SpecialtyToInstitutionOfEducationPostApiModel specialtyToIoE);
         Task DeleteSpecialtyToIoe(SpecialtyToInstitutionOfEducationPostApiModel specialtyToIoE);
     }
