+++ YIF_XUnitTests/Unit/YIF.Core.Service/Concrete/Services/IoEAdminServiceTests.cs
@@ -1,9 +1,13 @@
 ﻿using System;
+using System.Linq;
 using System.Resources;
 using System.Threading.Tasks;
 using System.Collections.Generic;
 using System.Linq.Expressions;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Configuration;
 using Moq;
+using AutoMapper;
 using SendGrid.Helpers.Errors.Model;
 using Xunit;
 using YIF.Core.Data.Entities;
@@ -11,6 +15,7 @@ using YIF.Core.Data.Interfaces;
 using YIF.Core.Domain.ApiModels.RequestApiModels;
 using YIF.Core.Domain.DtoModels.EntityDTO;
 using YIF.Core.Service.Concrete.Services;
+using YIF_XUnitTests.Unit.TestData;
 
 namespace YIF_XUnitTests.Unit.YIF.Core.Service.Concrete.Services
 {
@@ -18,22 +23,72 @@ namespace YIF_XUnitTests.Unit.YIF.Core.Service.Concrete.Services
     {
         private readonly IoEAdminService _ioEAdminService;
         private readonly Mock<ISpecialtyRepository<Specialty, SpecialtyDTO>> _specialtyRepository= new Mock<ISpecialtyRepository<Specialty, SpecialtyDTO>>();
+        private readonly Mock<IInstitutionOfEducationAdminRepository<InstitutionOfEducationAdminDTO>> _ioEAdminRepository =
+            new Mock<IInstitutionOfEducationAdminRepository<InstitutionOfEducationAdminDTO>>();
         private readonly Mock<ISpecialtyToInstitutionOfEducationRepository<SpecialtyToInstitutionOfEducation, SpecialtyToInstitutionOfEducationDTO>> 
             _specialtyToIoERepository = new Mock<ISpecialtyToInstitutionOfEducationRepository<SpecialtyToInstitutionOfEducation, SpecialtyToInstitutionOfEducationDTO>>();
         private readonly Mock<IInstitutionOfEducationRepository<InstitutionOfEducation, InstitutionOfEducationDTO>> 
             _ioERepository = new Mock<IInstitutionOfEducationRepository<InstitutionOfEducation, InstitutionOfEducationDTO>>();
         private readonly Mock<ResourceManager> _resourceManager = new Mock<ResourceManager>();
-        
+        private readonly Mock<IMapper> _mapper = new Mock<IMapper>();
+        private readonly Mock<IWebHostEnvironment> _env = new Mock<IWebHostEnvironment>();
+        private readonly Mock<IConfiguration> _configuration = new Mock<IConfiguration>();
+
         public IoEAdminServiceTests()
         {
             _ioEAdminService = new IoEAdminService(
                 _specialtyRepository.Object,
                 _ioERepository.Object,
                 _specialtyToIoERepository.Object,
+                _ioEAdminRepository.Object,
+                _mapper.Object,
+                _env.Object,
+                _configuration.Object,
                 _resourceManager.Object
             );
         }
 
+        [Fact]
+        public void ModifyDescriptionOfInstitution_WrongAdminId()
+        {
+            // Arrange
+            var wrongAdminId = "0";
+            var listOfAdmins = InstitutionOfEducationAdminTestData.GetIEnumerableInstitutionOfEducationAdminDTO();
+            _ioEAdminRepository.Setup(x => x.GetAllUniAdmins())
+                .Returns(Task.FromResult(listOfAdmins));
+
+            // Act
+            var result = _ioEAdminService.ModifyDescriptionOfInstitution(wrongAdminId, new InstitutionOfEducationPostApiModel());
+
+            // Assert
+            Assert.False(result.Result.Success);
+        }
+
+        [Fact]
+        public void ModifyDescriptionOfInstitution_ReturnTrue()
+        {
+            // Arrange
+            var listOfAdmins = InstitutionOfEducationAdminTestData.GetIEnumerableInstitutionOfEducationAdminDTO();
+            var institutionDTO = InstitutionOfEducationTestData.GetInstitutionOfEducationDTO();
+            var institution = InstitutionOfEducationTestData.GetInstitutionOfEducation();
+
+            _ioEAdminRepository.Setup(x => x.GetAllUniAdmins())
+                .Returns(Task.FromResult(listOfAdmins));
+            _mapper.Setup(x => x.Map<InstitutionOfEducationDTO>(It.IsAny<InstitutionOfEducationPostApiModel>()))
+                .Returns(institutionDTO);
+            _mapper.Setup(x => x.Map<InstitutionOfEducation>(It.IsAny<InstitutionOfEducationDTO>()))
+                .Returns(institution);
+            _ioERepository.Setup(x => x.Update(It.IsAny<InstitutionOfEducation>()))
+                .Returns(Task.FromResult(true));
+            _resourceManager.Setup(x => x.GetString(It.IsAny<string>()))
+                .Returns("");
+
+            // Act
+            var result = _ioEAdminService.ModifyDescriptionOfInstitution(listOfAdmins.FirstOrDefault().Id, new InstitutionOfEducationPostApiModel());
+
+            // Assert
+            Assert.True(result.Result.Success);
+        }
 
         [Fact]
         public async Task AddSpecialtyToIoE_ShouldAddSpecialty()
